static cast of final poses into float to solve narrowing issue

+++ components/graspDSR/src/specificworker.cpp
@@ -312,8 +312,8 @@ void SpecificWorker::inject_estimated_poses(RoboCompObjectPoseEstimationRGBD::Po
             }
 
             // inject estimated object pose into graph
-            vector<float> trans{final_pose->x(), final_pose->y(), final_pose->z()};
-            vector<float> rot{final_pose->rx(), final_pose->ry(), final_pose->rz()};
+            vector<float> trans{static_cast<float>(final_pose->x()), static_cast<float>(final_pose->y()), static_cast<float>(final_pose->z())};
+            vector<float> rot{static_cast<float>(final_pose->rx()), static_cast<float>(final_pose->ry()), static_cast<float>(final_pose->rz())};
             G->insert_or_assign_edge_RT(world.value(), id.value(), trans, rot);
 
             // ignore rest of objects
